+++ ClearentIdtechIOSFramework/ClearentCard.m
@@ -16,7 +16,7 @@
         self.softwareType = @"ClearentIdtechIOSFramework";
     }
     if(self.softwareTypeVersion == nil) {
-        self.softwareTypeVersion = @"v1";
+        self.softwareTypeVersion = @"v2.0.3";
     }
     
     if(self.csc != nil) {