+++ ClearentIdtechIOSFramework/ClearentDelegate.h
@@ -91,6 +91,44 @@ typedef enum {SWIPE=90} supportedNonEmvEntryMode;
 
 - (void) sendFeedback:(NSString*) message;
 
+/**
+Contactless Event
+During a Contactless transaction, if events are enabled, they will be sent to this protocol,
+
+@param event Event Type:
+- 01 = LED Event
+- 02 = Buzzer Event
+- 03 = LCD Message
+@param scheme LCD Message Scheme
+@param data Data
+   - When Event Type 01:
+   -- 0x00 = LED0 off
+   -- 0x10 = LED1 off
+   -- 0x20 = LED2 off
+   -- 0x30 = LED3 off
+   -- 0xF0 = ALL off
+   -- 0x01 = LED0 on
+   -- 0x11 = LED1 on
+   -- 0x21 = LED2 on
+   -- 0x31 = LED3 on
+   -- 0xF1 = ALL on
+   - When Event Type 02:
+   -- 0x10 = Short Beep No Change
+   -- 0x11 = Short Beep No Change
+   -- 0x12 = Double Short Beep
+   -- 0x13 = Triple Short Beep
+   -- 0x20 = 200ms Beep
+   -- 0x21 = 400ms Beep
+   -- 0x22 = 600ms Beep
+   - When Event Type 03:
+   -- Message ID (please refer to table in NEO Reference Guide)
+*/
+- (void) ctlsEvent:(Byte)event scheme:(Byte)scheme  data:(Byte)data;
+
+- (void) startFinalFeedbackMonitor:(int) timeout;
+
+- (void) disableCardRemovalTimer;
+
 @end