+++ ClearentIdtechIOSFramework/ClearentPublicVP3300Delegate.h
@@ -80,5 +80,4 @@
 
 - (void) bluetoothDevices:(NSArray<ClearentBluetoothDevice*>*) bluetoothDevices;
 
-
 @end