+++ ClearentIdtechIOSFramework/Clearent_VP3300.m
@@ -60,12 +60,13 @@
             clearentDelegate.runTransactionSelector = runTransactionSelector;
         }
         clearentDelegate = [[ClearentDelegate alloc] initWithPaymentCallback:publicDelegate clearentVP3300Configuration:clearentVP3300Configuration callbackObject:self withSelector:runTransactionSelector idTechSharedInstance: [IDT_VP3300 sharedController]];
+
+        [IDT_VP3300 sharedController].delegate = clearentDelegate;
         
         clearentDeviceConnector = [[ClearentDeviceConnector alloc] init:clearentDelegate clearentVP3300:self ];
         [clearentDelegate setClearentDeviceConnector:clearentDeviceConnector];
         clearentTransactions = [[ClearentTransactions alloc] init:clearentDelegate clearentVP3300:self];
 
-        [IDT_VP3300 sharedController].delegate = clearentDelegate;
         
         if(!clearentVP3300Configuration.disableRemoteLogging) {
             TELEPORT_DEBUG = YES;